+++ src/main/java/com/toy/attendance/dev/model/account/repository/AccountRepository.java
@@ -2,6 +2,7 @@ package com.toy.attendance.dev.model.account.repository;
 
 
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,6 +11,6 @@ import com.toy.attendance.dev.model.account.entity.Account;
 
 public interface AccountRepository extends JpaRepository<Account, Long>, AccountRepositoryCustom{
 
-    Account findByKakaoIdAndUseYn(Long kakaoId, String useYn);
+    Account findByKakaoIdAndUseYn(BigInteger kakaoId, String useYn);
     
 }
