+++ src/main/java/com/toy/attendance/dev/model/attendance/repository/AttendanceRepository.java
@@ -2,6 +2,7 @@ package com.toy.attendance.dev.model.attendance.repository;
 
 
 
+import java.math.BigInteger;
 import java.sql.Date;
 import java.util.List;
 
@@ -12,9 +13,9 @@ import com.toy.attendance.dev.model.attendance.entity.Attendance;
 
 public interface AttendanceRepository extends JpaRepository<Attendance, Long>, AttendanceRepositoryCustom{
 
-    Attendance findByAccountIdAndAttendanceDateAndUseYn(Long account_id, Date attendanceDate, String string);
+    Attendance findByAccountIdAndAttendanceDateAndUseYn(BigInteger account_id, Date attendanceDate, String string);
 
-    Attendance findByAttendanceIdAndUseYn(Long attendanceId, String string);
+    Attendance findByAttendanceIdAndUseYn(BigInteger attendanceId, String string);
 
     List<Attendance> findAllByUseYn(String string);
